backend/pyproject.toml
[project]
name = "berkeley-ai-25"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.9.4",
    "python-dotenv>=1.1.0",
]
